LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/cookiestore/change_eventhandler_for_already_expired.https.window-expected.txt


PASS CookieStore setting already-expired cookie should not be observed
FAIL CookieStore setting already-expired partitioned cookie should not be observed assert_equals: Deletion not observed after document.cookie sets already-expired cookie: number of changed cookies expected 1 but got 0
